Refactor fpga support code into fpgalib package.
Review: @DanielSank
