--- GHzDACs/Cleanup/todo.txt
@@ -1,8 +1,0 @@
-dac.py
- Remove buildParams
- Remove boardParams
- Refactor to FPGA class
-
-fpga server
- check invocations of dac.<method name>
- check invocations of adc.<method name>
