+++ fpgalib/TODO.txt
@@ -1,3 +1,12 @@
+dac.py
+ Remove buildParams
+ Remove boardParams
+ Refactor to FPGA class
+
+fpga server
+ check invocations of dac.<method name>
+ check invocations of adc.<method name>
+
 Immediate (yeah right):
 
 - Check SRAM start and end addresses via printout while running synchronized. Only two numbers should show up, one for each page
