+++ fpgalib/docs/HardRegProgram.txt
