+++ fpgalib/util.py
